+++ test/contract/execute-tool.test.js
@@ -1,3 +1,4 @@
-import { executeToolContract } from "./execute-tool.js";
+import { executeToolContract, executeToolMigrationContract } from "./execute-tool.js";
 
 executeToolContract("reference", () => async () => "contract result");
+executeToolMigrationContract("reference (migration)", () => async () => "contract result");