+++ test/contract/index.js
@@ -3,7 +3,6 @@
 export { transcriptStoreContract } from "./transcript-store.js";
 export { recallContract } from "./recall-contract.js";
 export { modelConfigProviderContract } from "./model-config-provider.js";
-export { executeToolContract } from "./execute-tool.js";
+export { executeToolContract, executeToolMigrationContract } from "./execute-tool.js";
 export { assemblyPortContract } from "./assembly-port.js";
-export { resourceStoreContract } from "./resource-store.js";
 export { notesStoreContract } from "./notes-store.js";